# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FinanIes",
        "contact": {
            "email": "derex@outlook.com.br"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/v1/company": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "get all company",
                "description": "get all company",
                "operationId": "e09a177de9205b901d0c546dafede205",
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Company"
                ],
                "summary": "register a new company",
                "description": "register a new company",
                "operationId": "d5834f135441d2027f02cd07d64e8b38",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "description",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/company/{company}": {
            "put": {
                "tags": [
                    "Company"
                ],
                "summary": "update a company by id",
                "description": "update a company by id",
                "operationId": "updateCompany",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "description": "company identification",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "description",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Company"
                ],
                "summary": "delete a company by id",
                "description": "delete a company by id",
                "operationId": "deleteCompany",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "description": "company identification",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/company/{company}/attach": {
            "post": {
                "tags": [
                    "Company"
                ],
                "summary": "associate permission to company",
                "description": "associate permission to company",
                "operationId": "dca1aec0c9ef1bfca32c0fc68e26b26c",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "description": "company identification",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "permission_id": {
                                        "description": "permission id",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "Bearer"
            }
        }
    }
}